restructure files for better readability and programmability
